requirements.txt
amqp==2.5.2
argcomplete==1.12.1
billiard==3.6.1.0
boto3==1.15.6
botocore==1.18.6
celery==4.3.0
certifi==2019.9.11
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Django==2.2
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-mysql==3.2.0
django-timezone-field==3.1
djangorestframework==3.10.3
dnspython==1.16.0
durationpy==0.5
eventlet==0.25.1
future==0.18.2
greenlet==0.4.15
gunicorn==20.0.2
hjson==3.0.2
idna==2.8
importlib-metadata==0.23
jmespath==0.10.0
kappa==0.6.0
kombu==4.6.6
monotonic==1.5
more-itertools==7.2.0
mysqlclient==1.4.5
pip-tools==5.3.1
placebo==0.9.0
PyJWT==1.7.1
python-crontab==2.4.0
python-dateutil==2.6.1
python-decouple==3.1
python-engineio==3.10.0
python-slugify==4.0.1
python-socketio==4.3.1
pytz==2019.3
PyYAML==5.3.1
raven==6.10.0
redis==3.3.11
requests==2.22.0
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.50.0
troposphere==2.6.2
twilio==6.33.0
urllib3==1.25.7
vine==1.3.0
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.50.0
zipp==0.6.0
